pyproject.toml
[project]
name = "abidpa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "chainlit>=2.6.9",
    "openai-agents>=0.2.7",
    "psycopg2>=2.9.10",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
]
